# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Your API",
    "description": "A simple example ASP.NET Core Web API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Article/type/{libelle}": {
      "get": {
        "tags": [
          "Article"
        ],
        "parameters": [
          {
            "name": "libelle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Distribution/{id}": {
      "get": {
        "tags": [
          "Distribution"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Distribution"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistributionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DistributionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DistributionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Distribution"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Distribution/statut/{statutLibelle}": {
      "get": {
        "tags": [
          "Distribution"
        ],
        "parameters": [
          {
            "name": "statutLibelle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DistributionSummaryDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DistributionSummaryDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DistributionSummaryDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/RefStatutDistribution": {
      "get": {
        "tags": [
          "RefStatutDistribution"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefStatutDistributionDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefStatutDistributionDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefStatutDistributionDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RefStatutDistribution"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefStatutDistributionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefStatutDistributionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefStatutDistributionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/RefStatutDistribution/{id}": {
      "get": {
        "tags": [
          "RefStatutDistribution"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RefStatutDistributionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefStatutDistributionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefStatutDistributionDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RefStatutDistribution"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefStatutDistributionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefStatutDistributionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefStatutDistributionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "RefStatutDistribution"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/RefType": {
      "get": {
        "tags": [
          "RefType"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefTypeDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefTypeDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefTypeDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RefType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/RefType/{id}": {
      "get": {
        "tags": [
          "RefType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RefTypeDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefTypeDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefTypeDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RefType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "RefType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArticleDTO": {
        "required": [
          "couleur",
          "description",
          "reference"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "prixDeVenteInter": {
            "type": "number",
            "format": "double"
          },
          "prixDeVenteRevendeur": {
            "type": "number",
            "format": "double"
          },
          "prixDeVenteClientFinal": {
            "type": "number",
            "format": "double"
          },
          "marqueId": {
            "type": "integer",
            "format": "int32"
          },
          "marqueLibelle": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "integer",
            "format": "int32"
          },
          "typeLibelle": {
            "type": "string",
            "nullable": true
          },
          "couleur": {
            "type": "string",
            "nullable": true
          },
          "fournisseurId": {
            "type": "integer",
            "format": "int32"
          },
          "fournisseurNom": {
            "type": "string",
            "nullable": true
          },
          "reference": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DistributionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "dateDistribution": {
            "type": "string",
            "format": "date-time"
          },
          "articleId": {
            "type": "integer",
            "format": "int32"
          },
          "societeId": {
            "type": "integer",
            "format": "int32"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "statutDistributionId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "DistributionSummaryDTO": {
        "type": "object",
        "properties": {
          "fournisseur": {
            "type": "string",
            "nullable": true
          },
          "client": {
            "type": "string",
            "nullable": true
          },
          "marque": {
            "type": "string",
            "nullable": true
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "dateDistribution": {
            "type": "string",
            "format": "date-time"
          },
          "quantite": {
            "type": "integer",
            "format": "int32"
          },
          "prixDeVente": {
            "type": "number",
            "format": "double"
          },
          "numFacture": {
            "type": "string",
            "nullable": true
          },
          "statut": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RefStatutDistributionDTO": {
        "required": [
          "code",
          "libelle"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "libelle": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RefTypeDTO": {
        "required": [
          "code",
          "libelle"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "libelle": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}